+++ src/App.tsx
@@ -1,5 +1,4 @@
 import './App.css';
-
 import Main from './pages/Main';
 
 function App() {
