+++ src/components/Header/Header.module.css
@@ -1,0 +1,13 @@
+.header {
+  padding: 20px 0;
+  line-height: 1.5em;
+}
+
+.title {
+  font-size: 6rem;
+  font-weight: 600;
+  margin-bottom: 2rem;
+  line-height: 1em;
+  color: #ececec;
+  text-align: center;
+}
