+++ src/components/InputTodo/InputTodo.module.css
@@ -1,0 +1,24 @@
+.form-container {
+  width: 100%;
+  margin-bottom: 20px;
+  display: flex;
+  border-radius: calc(0.5 * 100px);
+  box-shadow: 0 1px 6px 0 rgba(0, 0, 0, 0.38);
+  justify-content: space-evenly;
+}
+
+.input-text {
+  font-size: 1rem;
+  font-weight: 400;
+  width: 85%;
+  padding-right: 5px;
+  padding-left: 10px;
+  border-radius: calc(0.5 * 100px);
+  background-color: transparent;
+}
+
+.input-common {
+  height: 45px;
+  outline: none;
+  border: none;
+}
