+++ src/components/TodoItem/TodoItem.module.css
@@ -1,0 +1,30 @@
+.btn-trash {
+  color: orangered;
+  font-size: 16px;
+}
+
+.btn-trash:hover {
+  opacity: 0.5;
+}
+
+.item {
+  list-style-type: none;
+  padding: 17px 1.5rem;
+  border-bottom: 1px solid #eaeaea;
+  font-size: 1.2rem;
+  letter-spacing: 1.5px;
+}
+.item:hover {
+  opacity: 0.85;
+  background-color: #eaeaea;
+}
+
+.item-option {
+  float: right;
+}
+
+.item-option button {
+  background-color: transparent;
+  border: none;
+  cursor: pointer;
+}
