+++ src/components/TodoList/TodoList.module.css
@@ -1,0 +1,9 @@
+.empty-list {
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  font-size: 2.5rem;
+  letter-spacing: 1.5rem;
+  margin-left: 0.75rem;
+  color: #ececec;
+}
