+++ src/index.tsx
@@ -1,10 +1,10 @@
-import App from "./App";
-import ReactDOM from "react-dom";
-import { StrictMode } from "react";
+import ReactDOM from 'react-dom';
+import { StrictMode } from 'react';
+import App from './App';
 
 ReactDOM.render(
   <StrictMode>
     <App />
   </StrictMode>,
-  document.getElementById("root"),
+  document.getElementById('root'),
 );
